# Code/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CGDA Backend",
    "description": "Minimal backend service for the geographic analysis platform.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/demo/layers/snapshots": {
      "get": {
        "tags": [
          "demo"
        ],
        "summary": "List Demo Snapshots",
        "operationId": "list_demo_snapshots_demo_layers_snapshots_get",
        "parameters": [
          {
            "name": "hour",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 12,
              "title": "Hour"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoLayerSnapshotsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/demo/layers/{layer_id}/snapshot": {
      "get": {
        "tags": [
          "demo"
        ],
        "summary": "Get Demo Snapshot",
        "operationId": "get_demo_snapshot_demo_layers__layer_id__snapshot_get",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          },
          {
            "name": "hour",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 12,
              "title": "Hour"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoLayerSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/geo/transform": {
      "get": {
        "tags": [
          "geo"
        ],
        "summary": "Transform Geo Point",
        "operationId": "transform_geo_point_geo_transform_get",
        "parameters": [
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Lng"
            }
          },
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Lat"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source"
            }
          },
          {
            "name": "target",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "EPSG:3857",
              "title": "Target"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "number"
                      },
                      {
                        "type": "string"
                      }
                    ]
                  },
                  "title": "Response Transform Geo Point Geo Transform Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/overlay-preview/{layer_id}": {
      "get": {
        "tags": [
          "overlay"
        ],
        "summary": "Get Overlay Preview",
        "description": "返回图层的 PNG 预览图（地理配准），供前端 MapLibre image source 使用。\n\n对于时间序列图层，可通过 `?time=YYYYMMDD` 指定时间标签；\n未指定时使用 default_time。",
        "operationId": "get_overlay_preview_overlay_preview__layer_id__get",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/overlay-bounds/{layer_id}": {
      "get": {
        "tags": [
          "overlay"
        ],
        "summary": "Get Overlay Bounds",
        "description": "返回图层的地理边界信息 + 元数据，供前端 MapLibre image source 定位与时间控制使用。",
        "operationId": "get_overlay_bounds_overlay_bounds__layer_id__get",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Overlay Bounds Overlay Bounds  Layer Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/overlay-value/{layer_id}": {
      "get": {
        "tags": [
          "overlay"
        ],
        "summary": "Get Overlay Value",
        "description": "查询 overlay 图层在指定点 (lng, lat) 的像素值。\n\n对于时间序列图层，可通过 ?time=YYYYMMDD 指定时间标签。\n返回 {\"value\": float | null, \"unit\": str, \"layer_id\": str, ...}。",
        "operationId": "get_overlay_value_overlay_value__layer_id__get",
        "parameters": [
          {
            "name": "layer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Layer Id"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Lng"
            }
          },
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Lat"
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Overlay Value Overlay Value  Layer Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DemoAvailabilityState": {
        "type": "string",
        "enum": [
          "empty",
          "partial",
          "ready"
        ],
        "title": "DemoAvailabilityState"
      },
      "DemoDataStateMode": {
        "type": "string",
        "enum": [
          "demo",
          "placeholder",
          "mixed"
        ],
        "title": "DemoDataStateMode"
      },
      "DemoFieldAliasMap": {
        "properties": {
          "metric_value": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Metric Value"
          },
          "hotspot_value": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hotspot Value"
          },
          "observation_time": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Observation Time"
          },
          "status_label": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Status Label"
          }
        },
        "type": "object",
        "required": [
          "metric_value",
          "hotspot_value",
          "observation_time",
          "status_label"
        ],
        "title": "DemoFieldAliasMap"
      },
      "DemoLayerSnapshot": {
        "properties": {
          "layer_id": {
            "type": "string",
            "title": "Layer Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "metric_label": {
            "type": "string",
            "title": "Metric Label"
          },
          "metric_unit": {
            "type": "string",
            "title": "Metric Unit"
          },
          "metric_precision": {
            "type": "integer",
            "title": "Metric Precision"
          },
          "update_label": {
            "type": "string",
            "title": "Update Label"
          },
          "source_label": {
            "type": "string",
            "title": "Source Label"
          },
          "accent_color": {
            "type": "string",
            "title": "Accent Color"
          },
          "accent_glow": {
            "type": "string",
            "title": "Accent Glow"
          },
          "chip_tone": {
            "type": "string",
            "title": "Chip Tone"
          },
          "data_state_mode": {
            "$ref": "#/components/schemas/DemoDataStateMode"
          },
          "data_state_label": {
            "type": "string",
            "title": "Data State Label"
          },
          "empty_state_label": {
            "type": "string",
            "title": "Empty State Label"
          },
          "availability_state": {
            "$ref": "#/components/schemas/DemoAvailabilityState"
          },
          "trend_label": {
            "type": "string",
            "title": "Trend Label"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "status_label": {
            "type": "string",
            "title": "Status Label"
          },
          "confidence_label": {
            "type": "string",
            "title": "Confidence Label"
          },
          "requested_hour": {
            "type": "number",
            "title": "Requested Hour"
          },
          "field_aliases": {
            "$ref": "#/components/schemas/DemoFieldAliasMap"
          },
          "raw_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Payload"
          }
        },
        "type": "object",
        "required": [
          "layer_id",
          "display_name",
          "category",
          "metric_label",
          "metric_unit",
          "metric_precision",
          "update_label",
          "source_label",
          "accent_color",
          "accent_glow",
          "chip_tone",
          "data_state_mode",
          "data_state_label",
          "empty_state_label",
          "availability_state",
          "trend_label",
          "summary",
          "status_label",
          "confidence_label",
          "requested_hour",
          "field_aliases",
          "raw_payload"
        ],
        "title": "DemoLayerSnapshot"
      },
      "DemoLayerSnapshotsResponse": {
        "properties": {
          "requested_hour": {
            "type": "number",
            "title": "Requested Hour"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DemoLayerSnapshot"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "requested_hour",
          "items"
        ],
        "title": "DemoLayerSnapshotsResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}